requirements.txt
streamlit==1.29.0
pandas==2.2.2
openai==1.40.6
python-dotenv==1.0.1
openpyxl==3.1.5
